pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sycop"
version = "0.1.0"
description = "Sycophancy Under Pressure: trajectory-level evaluation of policy-compliant manipulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
]

[project.scripts]
sycop = "sycop.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sycop*"]

